refactor: move test fixtures to solver conftest
